include , exclude

+++ DP/Coin_Change.cpp
@@ -1,6 +1,6 @@
 class Solution {
   public:
-    long long int solve(int coins[], int N, int sum,int idx,vector<vector<int>>&dp){
+    long long int solve(int coins[], int N, int sum,int idx,vector<vector<long long int>>&dp){
         // base 
         if(sum == 0){
             return 1;
@@ -11,13 +11,13 @@ class Solution {
         if(dp[sum][idx] != -1){
             return dp[sum][idx];
         }
-        int include = solve(coins,N,sum-coins[idx],idx,dp);
-        int exclude = solve(coins,N,sum,idx+1,dp);
+        long long int include = solve(coins,N,sum-coins[idx],idx,dp);
+        long long int exclude = solve(coins,N,sum,idx+1,dp);
         return dp[sum][idx] = include + exclude;
     }
     long long int count(int coins[], int N, int sum) {
 
-        vector<vector<int>>dp(sum+1,vector<int>(N+1,-1));
+        vector<vector<long long int>>dp((long long int)sum+1,vector<long long int>(N+1,-1));
         return solve(coins,N,sum,0,dp);
         
     }
